+++ pathspec/patterns/__init__.py
@@ -1,0 +1,8 @@
+# encoding: utf-8
+"""
+The *pathspec.patterns* package contains the pattern matching
+implementations.
+"""
+
+# Load pattern implementations.
+from . import gitwildmatch